+++ lib/model/Assistant.dart
@@ -82,40 +82,8 @@ class Message {
     this.sourcesExpanded, this.feedbackResponseType, this.isNegativeFeedbackMessage});
 
   factory Message.fromAnswerJson(Map<String, dynamic> json) {
-    List<String>? sources = JsonUtils.stringListValue(json['sources']);
-    if (sources == null) {
-      String? source = JsonUtils.stringValue(json['sources']);
-      if (StringUtils.isNotEmpty(source)) {
-        sources = source!.split((RegExp(r'[,\n]')));
-        sources = sources.map((e) => e.trim()).toList();
-      }
-    }
-
-    List<Link>? deeplinks = Link.listFromJson(json['deeplinks']);
-    String? deeplink = JsonUtils.stringValue(json['deeplink'])?.trim();
-    if (deeplink != null) {
-      if (deeplinks == null) {
-        deeplinks = [];
-      }
-      deeplinks.add(Link(name: deeplinkNameMap[deeplink] ?? deeplink.split('.|_').join(' '), link: deeplink));
-    }
-
-    return Message(
-      id: JsonUtils.stringValue(json['id'])?.trim() ?? '',
-      content: JsonUtils.stringValue(json['answer'])?.trim() ?? '',
-      user: JsonUtils.boolValue(json['user']) ?? false,
-      example: JsonUtils.boolValue(json['example']) ?? false,
-      queryLimit: JsonUtils.intValue(json['query_limit']),
-      acceptsFeedback: JsonUtils.boolValue(json['accepts_feedback']) ?? true,
-      links: deeplinks,
-      sources: sources ?? [],
-      feedback: null,
-      feedbackExplanation: null,
-    );
-  }
-
-  factory Message.fromJson(Map<String, dynamic> json) {
     Map<String, dynamic>? answerJson = JsonUtils.mapValue(json['answer']);
+    Map<String, dynamic>? feedbackJson = JsonUtils.mapValue(json['feedback']);
 
     List<String>? sources = JsonUtils.stringListValue(answerJson?['sources']);
     if (sources == null) {
@@ -144,9 +112,33 @@ class Message {
       acceptsFeedback: JsonUtils.boolValue(answerJson?['accepts_feedback']) ?? true,
       links: deeplinks,
       sources: sources ?? [],
-      feedback: null,
-      feedbackExplanation: null,
+      feedback: _feedbackFromString(JsonUtils.stringValue(feedbackJson?['feedback'])),
+      feedbackExplanation: JsonUtils.stringValue(feedbackJson?['explanation']),
     );
+  }
+
+  factory Message.fromQueryJson(Map<String, dynamic> json) {
+    Map<String, dynamic>? queryJson = JsonUtils.mapValue(json['query']);
+
+    return Message(
+        id: JsonUtils.stringValue(json['id'])?.trim() ?? '',
+        content: StringUtils.ensureNotEmpty(JsonUtils.stringValue(queryJson?['question'])),
+        user: true,
+        example: false,
+        acceptsFeedback: false,
+        feedback: null,
+        feedbackExplanation: null);
+  }
+
+  static MessageFeedback? _feedbackFromString(String? value) {
+    switch (value) {
+      case 'bad':
+        return MessageFeedback.bad;
+      case 'good':
+        return MessageFeedback.good;
+      default:
+        return null;
+    }
   }
 
   bool get isAnswerUnknown => (content.toLowerCase() == _unknownAnswerValue.toLowerCase());