+++ lib/ui/assistant/AssistantConversationContentWidget.dart
@@ -88,6 +88,10 @@ class _AssistantConversationContentWidgetState extends State<AssistantConversati
     _onPullToRefresh();
 
     _userContext = _getUserContext();
+
+    if (CollectionUtils.isNotEmpty(Assistant().messages)) {
+      _shouldScrollToBottom = true;
+    }
   }
 
   @override
@@ -317,7 +321,12 @@ class _AssistantConversationContentWidgetState extends State<AssistantConversati
   void _onTapSourcesAndLinksLabel(Message message) {
     setStateIfMounted(() {
       message.sourcesExpanded = !(message.sourcesExpanded ?? false);
-      _shouldScrollToBottom = true;
+      int msgsLength = Assistant().messages.length;
+      int msgIndex = (msgsLength > 0) ? Assistant().messages.indexOf(message) : -1;
+      if ((msgIndex >= 0) && (msgIndex == (msgsLength - 1))) {
+        // Automatically scroll only if the last message "Sources and Links" label is tapped
+        _shouldScrollToBottom = true;
+      }
     });
   }
 
@@ -896,10 +905,9 @@ class _AssistantConversationContentWidgetState extends State<AssistantConversati
   }
 
   void _clearAllMessages() {
-    setStateIfMounted(() {
-      Assistant().clearMessages();
-      _shouldScrollToBottom = true;
-    });
+    //TBD: DD - implement when we have a backend API
+    AppAlert.showMessage(context, 'Not implemented, yet.');
+    Assistant().removeAllMessages();
   }
 
   void _scrollToBottomIfNeeded() {