+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Focus on "Sources and Links" in Assistant [#4128](https://github.com/rokwire/illinois-app/issues/4128).
+- Keep Illinois Assistant history until user clears it [#4129](https://github.com/rokwire/illinois-app/issues/4129).
+### Changed
+- Update text styles used for Events card  and Detail Panel[#4158]
+
 ## [6.0.12] - 2024-06-24
 ### Changed
 - Use static and permanently visible gradient shadow in HomePanel top [#4108](https://github.com/rokwire/illinois-app/issues/4108).